requirements.txt
typer[all]
rich
toml
requests
openai>=1.0.0
python-dotenv
